Refactor
https://coveord.atlassian.net/browse/KIT-1411
