+++ packages/atomic/src/components/facets/facet-common.ts
@@ -1,27 +1,10 @@
-import {
-  CategoryFacetValue,
-  DateFacetValue,
-  FacetValue,
-  loadFacetSetActions,
-  loadSearchActions,
-  loadSearchAnalyticsActions,
-  NumericFacetValue,
-  SearchEngine,
-  SearchStatus,
-  SearchStatusState,
-} from '@coveo/headless';
-import {ObservableMap} from '@stencil/store';
+import {SearchStatus, SearchStatusState} from '@coveo/headless';
 import {i18n} from 'i18next';
-import {groupBy} from 'lodash';
+import {DependsOnParam} from '../../utils/store';
 import {
-  FacetRequestWithType,
-  AtomicStore,
-  DependsOnParam,
-  DependencyCondition,
-  getFacet,
-  DependantFacet,
-  FacetWithType,
-} from '../../utils/store';
+  buildFacetDependsOnAnyValueCondition,
+  buildFacetDependsOnSpecificValueCondition,
+} from './facet-depends-on/facet-dependency-conditions';
 
 export interface BaseFacet<Facet, FacetState> {
   facet?: Facet;
@@ -46,38 +29,6 @@ export interface FacetValueProps {
   buttonRef?: (element?: HTMLButtonElement) => void;
 }
 
-export interface IncorrectFacetEnableState {
-  facetWithType: FacetWithType;
-  shouldBeEnabled: boolean;
-}
-
-export const facetHasSelectedValue: DependencyCondition = ({request}) => {
-  return !!(
-    request.currentValues as (
-      | FacetValue
-      | CategoryFacetValue
-      | NumericFacetValue
-      | DateFacetValue
-    )[]
-  ).find((value) => value.state === 'selected');
-};
-
-export function buildFacetDependsOnValueCondition(
-  expectedValue: string
-): DependencyCondition {
-  return (args) => {
-    if (!facetHasSelectedValue(args)) {
-      return false;
-    }
-    if (args.type === 'facets' || args.type === 'categoryFacets') {
-      return args.request.currentValues.some(
-        (value) => value.state === 'selected' && expectedValue === value.value
-      );
-    }
-    return true;
-  };
-}
-
 export function parseDependsOn(
   dependsOn: Record<string, string>
 ): DependsOnParam[] {
@@ -85,103 +36,13 @@ export function parseDependsOn(
     if (!value) {
       return {
         parentFacetId: facetId,
-        isDependencyMet: facetHasSelectedValue,
+        isDependencyMet: buildFacetDependsOnAnyValueCondition(),
       };
     } else {
       return {
         parentFacetId: facetId,
-        isDependencyMet: buildFacetDependsOnValueCondition(value),
+        isDependencyMet: buildFacetDependsOnSpecificValueCondition(value),
       };
     }
   });
-}
-
-function checkIfAnyDependencyMet(
-  engine: SearchEngine,
-  state: AtomicStore,
-  dependencies: DependantFacet[]
-) {
-  return dependencies.some((dependency) => {
-    const facetWithType = getFacet(engine, state, dependency.parentFacetId);
-    if (!facetWithType) {
-      return false;
-    }
-    return dependency.isDependencyMet({
-      type: facetWithType.type,
-      request: facetWithType.request,
-    } as FacetRequestWithType);
-  });
-}
-
-function thawAllFacets(engine: SearchEngine) {
-  Object.keys(engine.state.facetSet ?? {}).forEach((facetId) =>
-    engine.dispatch(
-      loadFacetSetActions(engine).updateFreezeCurrentValues({
-        facetId,
-        freezeCurrentValues: false,
-      })
-    )
-  );
-}
-
-export function getIncorrectFacetEnableStates(
-  engine: SearchEngine,
-  state: AtomicStore
-) {
-  const dependenciesPerDependantFacet = Object.entries(
-    groupBy(
-      state.facetDependencies,
-      (dependency) => dependency.dependantFacetId
-    )
-  );
-  const incorrectFacetStates: IncorrectFacetEnableState[] = [];
-  dependenciesPerDependantFacet.forEach(([dependantFacetId, dependencies]) => {
-    const areDependenciesMet = checkIfAnyDependencyMet(
-      engine,
-      state,
-      dependencies
-    );
-    const facetWithType = getFacet(engine, state, dependantFacetId);
-    if (!facetWithType) {
-      return;
-    }
-    const isFacetEnabled = facetWithType.facet.state.enabled;
-    if (areDependenciesMet !== isFacetEnabled) {
-      incorrectFacetStates.push({
-        facetWithType: facetWithType,
-        shouldBeEnabled: areDependenciesMet,
-      });
-    }
-  });
-  return incorrectFacetStates;
-}
-
-export function fixFacetDependencies(
-  engine: SearchEngine,
-  store: ObservableMap<AtomicStore>
-) {
-  const state = store.state;
-  let facetStatesToUpdate = getIncorrectFacetEnableStates(engine, state);
-  let facetWasEnabled = false;
-  let facetValuesWereCleared = false;
-  while (facetStatesToUpdate.length) {
-    facetStatesToUpdate.forEach(({facetWithType, shouldBeEnabled}) => {
-      if (shouldBeEnabled) {
-        facetWithType.facet.enable();
-        facetWasEnabled = true;
-      } else {
-        facetWithType.facet.disable();
-        facetValuesWereCleared ||= facetHasSelectedValue(facetWithType);
-      }
-      facetStatesToUpdate = getIncorrectFacetEnableStates(engine, state);
-    });
-  }
-  if (facetWasEnabled || facetValuesWereCleared) {
-    thawAllFacets(engine);
-    engine.dispatch(
-      loadSearchActions(engine).executeSearch(
-        loadSearchAnalyticsActions(engine).logClearBreadcrumbs()
-      )
-    );
-  }
 }
